requirements.txt
# ASIMOD Core Dependencies
llama-cpp-python>=0.2.0
pytest>=7.0.0
requests>=2.28.0
python-dotenv>=1.0.0
numpy>=1.23.0
psutil>=5.9.0

# API and Web Server
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6

# UI and Media (Sistema Module)
opencv-python>=4.7.0.0
pygame>=2.1.0
Pillow>=9.0.0
pymupdf>=1.20.0

# Voice and Audio (STT/TTS)
SpeechRecognition>=3.10.0
PyAudio>=0.2.13

# LLM Adapters
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0
